+++ resources/js/Components/Global/Schedule/ScheduleGridContainer.vue
@@ -348,7 +348,8 @@ function handleShowClick(item) {
 
   switch (item.type) {
     case 'show':
-      url = `/shows/${item.content.slug}/`
+      // url = `/shows/${item.content.slug}/`
+      url = `/teams/${item.content.teamSlug}/`
       break
     case 'movie':
       url = `/movies/${item.content.slug}/`
@@ -359,9 +360,7 @@ function handleShowClick(item) {
     default:
       // Handle default case or do nothing
   }
-
   Inertia.visit(url) // Visit the dynamically created URL
-
 }