+++ resources/js/Stores/UserStore.js
@@ -199,6 +199,13 @@ export const useUserStore = defineStore('userStore', {
             }
             return dayjs.utc(dateTime).tz(this.timezone).format(formatString)
         },
+        formatLongDateTimeFromUtcToUserTimezone(dateTime, formatString = 'dddd MMMM D [at] h:mm a') {
+            if (!this.timezone) {
+                console.error('Timezone is not set.')
+                return dateTime // Or handle this case as appropriate for your app
+            }
+            return dayjs.utc(dateTime).tz(this.timezone).format(formatString)
+        },
         formatTimeInUserTimezone(time, formatString = 'h:mm a') {
             if (!this.timezone) {
                 console.error('Timezone is not set.')